# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NABU VPN indexer API",
        "title": "vpn-indexer",
        "contact": {
            "name": "Blink Labs Software",
            "url": "https://blinklabs.io",
            "email": "support@blinklabs.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/client/available": {
            "post": {
                "description": "Check if a client profile is available",
                "consumes": [
                    "application/json"
                ],
                "summary": "ClientAvailable",
                "parameters": [
                    {
                        "description": "Client Available Request",
                        "name": "ClientAvailableRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ClientAvailableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/client/list": {
            "post": {
                "description": "Search for clients matching a given manager public key hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ClientList",
                "parameters": [
                    {
                        "description": "List Request",
                        "name": "ClientListRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ClientListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching clients",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Client"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/client/profile": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch a client VPN profile via a signed S3 link; authenticate with a session Bearer token and provide the subscription id in the body",
                "consumes": [
                    "application/json"
                ],
                "summary": "ClientProfile",
                "parameters": [
                    {
                        "description": "Profile Request",
                        "name": "ClientProfileRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ClientProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Client": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                }
            }
        },
        "api.ClientAvailableRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "api.ClientListRequest": {
            "type": "object",
            "properties": {
                "ownerAddress": {
                    "type": "string"
                }
            }
        },
        "api.ClientProfileRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Session token from POST /api/auth/session, sent as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}